Restructure confirmation footer as dispatcher plus variants
